rangement routes equivalence
