+++ api/package.json
@@ -39,6 +39,7 @@
     "@godaddy/terminus": "^4.12.1",
     "@sentry/node": "^8.25.0",
     "@sentry/profiling-node": "^8.25.0",
+    "@types/passport": "^1.0.17",
     "archiver": "^7.0.1",
     "aws-sdk": "^2.1248.0",
     "bcryptjs": "^2.4.3",
